backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
alembic>=1.13.1
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
apscheduler>=3.10.0
yfinance>=0.2.38
httpx>=0.27.0
python-multipart>=0.0.9
